requirements.txt
langchain
langchain-community
langchain_groq
langchain_huggingface
streamlit
python-dotenv
sentence-transformers
pandas
chromadb
